Add ignoreEINTR function for CPointer
