+++ library/file/src/nativeMain/kotlin/io/matthewnelson/kmp/file/internal/NativePlatform.kt
@@ -13,21 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-@file:Suppress("KotlinRedundantDiagnosticSuppress", "NOTHING_TO_INLINE", "VariableInitializerIsRedundant")
+@file:Suppress("KotlinRedundantDiagnosticSuppress", "NOTHING_TO_INLINE")
 
 package io.matthewnelson.kmp.file.internal
 
 import io.matthewnelson.kmp.file.*
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.memScoped
-import kotlinx.cinterop.toKString
-import platform.posix.EINTR
-import platform.posix.EINVAL
 import platform.posix.errno
-import platform.posix.strerror
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.InvocationKind
-import kotlin.contracts.contract
 
 @Throws(IOException::class)
 @OptIn(DelicateFileApi::class, ExperimentalForeignApi::class)
@@ -99,36 +92,4 @@ internal actual inline fun File.platformWriteUtf8(text: String) {
     }
 
     platformWriteBytes(encoded)
-}
-
-@ExperimentalForeignApi
-internal inline fun errnoToString(errno: Int): String {
-    return strerror(errno)?.toKString() ?: "errno: $errno"
-}
-
-@ExperimentalForeignApi
-internal inline fun errnoToIllegalArgumentOrIOException(errno: Int): Exception {
-    return if (errno == EINVAL) {
-        val message = errnoToString(errno)
-        IllegalArgumentException(message)
-    } else {
-        errnoToIOException(errno)
-    }
-}
-
-// action must be something that returns -1 and sets errno
-@OptIn(ExperimentalContracts::class)
-internal inline fun ignoreEINTR(action: () -> Int): Int {
-    contract {
-        callsInPlace(action, InvocationKind.UNKNOWN)
-    }
-
-    var ret = -1
-    while (true) {
-        ret = action()
-        if (ret != -1) break
-        if (errno == EINTR) continue
-        break
-    }
-    return ret
 }
